requirements.txt
ollama
requests
fastapi
uvicorn
pydantic
chromadb
sentence-transformers
opencv-python
pillow
numpy
asyncio
json
